refactor: move remote loading to its own file
